Help command fixed

+++ Events/automod.js
@@ -14,11 +14,11 @@ module.exports = {
                     message.channel.send(client.embed("Warning", `${message.author.tag} swearing is not allowed in ${message.guild.name}.`)).catch(e => {return;})
                 }
                }
-               
+               if(d.features.includes("antilink")){
                     if(message.content.toLowerCase().replace(" ", "").includes("discord.gg/")){
                      message.delete().catch(e => {return;})
                     message.channel.send(client.embed("Warning", `${message.author.tag} sending links is not allowed in ${message.guild.name}`)).catch(e => {return;})
-                    
+                    }
                 }
                }
            }
